test(api): split test_api.py into per-group modules + migrate call sites (#144)
Companion to the api split. unit_test/test_api.py (1426 lines) is split
into one file per sub-API under unit_test/github/:

  test_repo_api.py    (17 tests)
  test_org_api.py     (12 tests)
  test_commit_api.py  (9 tests)
  test_action_api.py  (10 tests)
  test_user_api.py    (7 tests)

GitHub App auth is exercised indirectly via test_app_enum.py, so
no separate test_app_api.py is needed.

The original test_api.py shrinks to lifecycle / rate-limit / proxy /
session-bootstrap tests that exercise the parent Api orchestrator
itself, since those do not fit into a sub-API group.

A small unit_test/api_mock.py helper was extracted so each per-group
test file can import the same Api fixture instead of duplicating the
mock-construction boilerplate.

Other test modules (test_attack, test_oidc_attack, test_runner_attack,
test_secrets_attack, test_enumerate, test_app_enum, test_org_enumerate,
test_repo_enumerate, test_search, test_cli, test_finegrained_enumeration_simple,
graph/*, persistence/*) got mechanical updates to match the new
self.api.<group>.method call sites.

Verified:
  pytest -q  -> 556 passed
  ruff       -> clean
