+++ unit_test/graph/test_artifact_poisoning.py
@@ -9,12 +9,12 @@ import pytest
 from gatox.enumerate.results.complexity import Complexity
 from gatox.enumerate.results.confidence import Confidence
 from gatox.enumerate.results.issue_type import IssueType
-from gatox.github.api import Api
 from gatox.workflow_graph.graph.tagged_graph import TaggedGraph
 from gatox.workflow_graph.visitors.artifact_poisoning_visitor import (
     ArtifactPoisoningVisitor,
 )
 from gatox.workflow_graph.visitors.visitor_utils import VisitorUtils
+from unit_test.api_mock import make_api_mock
 
 
 @pytest.fixture
@@ -26,7 +26,7 @@ def mock_graph():
 @pytest.fixture
 def mock_api():
     """Mock Api fixture"""
-    return MagicMock(spec=Api)
+    return make_api_mock()
 
 
 @pytest.fixture
