+++ unit_test/persistence/test_persistence_api.py
@@ -17,7 +17,7 @@ async def test_invite_collaborator_success(api):
     api.call_put = AsyncMock()
     api.call_put.return_value.status_code = 201
 
-    result = await api.invite_collaborator("test/repo", "username")
+    result = await api.repo.invite_collaborator("test/repo", "username")
 
     assert result is True
     api.call_put.assert_called_once_with(
@@ -31,7 +31,7 @@ async def test_invite_collaborator_already_invited(api):
     api.call_put = AsyncMock()
     api.call_put.return_value.status_code = 204
 
-    result = await api.invite_collaborator("test/repo", "username")
+    result = await api.repo.invite_collaborator("test/repo", "username")
 
     assert result is True
 
@@ -42,7 +42,7 @@ async def test_invite_collaborator_failure(api):
     api.call_put = AsyncMock()
     api.call_put.return_value.status_code = 403
 
-    result = await api.invite_collaborator("test/repo", "username")
+    result = await api.repo.invite_collaborator("test/repo", "username")
 
     assert result is False
 
@@ -53,7 +53,7 @@ async def test_create_deploy_key_success(api):
     api.call_post = AsyncMock()
     api.call_post.return_value.status_code = 201
 
-    result = await api.create_deploy_key(
+    result = await api.repo.create_deploy_key(
         "test/repo", "Test Key", "ssh-rsa AAAA...", read_only=False
     )
 
@@ -70,7 +70,7 @@ async def test_create_deploy_key_failure(api):
     api.call_post = AsyncMock()
     api.call_post.return_value.status_code = 422
 
-    result = await api.create_deploy_key("test/repo", "Test Key", "invalid-key")
+    result = await api.repo.create_deploy_key("test/repo", "Test Key", "invalid-key")
 
     assert result is False
 
@@ -81,7 +81,7 @@ async def test_create_workflow_on_branch_success(api):
     from unittest.mock import MagicMock
 
     # Mock the repository info call
-    api.get_repository = AsyncMock(return_value={"default_branch": "main"})
+    api.repo.get_repository = AsyncMock(return_value={"default_branch": "main"})
 
     # Mock the branch info call
     branch_response = MagicMock()
@@ -93,25 +93,25 @@ async def test_create_workflow_on_branch_success(api):
     create_branch_response.status_code = 201
 
     # Mock the commit_file call
-    api.commit_file = AsyncMock(return_value="def456")
+    api.commit.commit_file = AsyncMock(return_value="def456")
 
     api.call_get = AsyncMock(return_value=branch_response)
     api.call_post = AsyncMock(return_value=create_branch_response)
 
-    result = await api.create_workflow_on_branch(
+    result = await api.commit.create_workflow_on_branch(
         "test/repo", "feature-branch", "test.yml", "workflow content"
     )
 
     assert result == "def456"
 
     # Verify all expected calls were made
-    api.get_repository.assert_called_once_with("test/repo")
+    api.repo.get_repository.assert_called_once_with("test/repo")
     api.call_get.assert_called_once_with("/repos/test/repo/git/ref/heads/main")
     api.call_post.assert_called_once_with(
         "/repos/test/repo/git/refs",
         params={"ref": "refs/heads/feature-branch", "sha": "abc123"},
     )
-    api.commit_file.assert_called_once_with(
+    api.commit.commit_file.assert_called_once_with(
         repo_name="test/repo",
         branch_name="feature-branch",
         file_path=".github/workflows/test.yml",
@@ -125,9 +125,9 @@ async def test_create_workflow_on_branch_success(api):
 @pytest.mark.asyncio
 async def test_create_workflow_on_branch_repo_failure(api):
     """Test workflow creation failure due to repository access."""
-    api.get_repository = AsyncMock(return_value=None)
+    api.repo.get_repository = AsyncMock(return_value=None)
 
-    result = await api.create_workflow_on_branch(
+    result = await api.commit.create_workflow_on_branch(
         "test/repo", "feature-branch", "test.yml", "workflow content"
     )
