+++ unit_test/persistence/test_persistence_attack.py
@@ -17,14 +17,14 @@ async def test_invite_collaborators_success(mock_output, persistence_attacker):
     """Test successful collaborator invitation."""
     # Mock the setup and API calls
     persistence_attacker.setup_user_info = AsyncMock(return_value=True)
-    persistence_attacker.api.invite_collaborator = AsyncMock(return_value=True)
+    persistence_attacker.api.repo.invite_collaborator = AsyncMock(return_value=True)
 
     result = await persistence_attacker.invite_collaborators(
         "test/repo", ["user1", "user2"], "push"
     )
 
     assert result is True
-    assert persistence_attacker.api.invite_collaborator.call_count == 2
+    assert persistence_attacker.api.repo.invite_collaborator.call_count == 2
 
 
 @pytest.mark.asyncio
@@ -34,7 +34,9 @@ async def test_invite_collaborators_partial_success(mock_output, persistence_att
     persistence_attacker.setup_user_info = AsyncMock(return_value=True)
 
     # Mock first invite success, second failure
-    persistence_attacker.api.invite_collaborator = AsyncMock(side_effect=[True, False])
+    persistence_attacker.api.repo.invite_collaborator = AsyncMock(
+        side_effect=[True, False]
+    )
 
     result = await persistence_attacker.invite_collaborators(
         "test/repo", ["user1", "user2"], "push"
@@ -62,14 +64,14 @@ async def test_invite_collaborators_setup_failure(mock_output, persistence_attac
 async def test_create_deploy_key_success(mock_output, persistence_attacker):
     """Test successful deploy key creation."""
     persistence_attacker.setup_user_info = AsyncMock(return_value=True)
-    persistence_attacker.api.create_deploy_key = AsyncMock(return_value=True)
+    persistence_attacker.api.repo.create_deploy_key = AsyncMock(return_value=True)
 
     result = await persistence_attacker.create_deploy_key(
         "test/repo", "Test Key", "/tmp/test_key.pem"
     )
 
     assert result is True
-    persistence_attacker.api.create_deploy_key.assert_called_once()
+    persistence_attacker.api.repo.create_deploy_key.assert_called_once()
 
 
 @pytest.mark.asyncio
@@ -77,7 +79,7 @@ async def test_create_deploy_key_success(mock_output, persistence_attacker):
 async def test_create_deploy_key_failure(mock_output, persistence_attacker):
     """Test deploy key creation failure."""
     persistence_attacker.setup_user_info = AsyncMock(return_value=True)
-    persistence_attacker.api.create_deploy_key = AsyncMock(return_value=False)
+    persistence_attacker.api.repo.create_deploy_key = AsyncMock(return_value=False)
 
     result = await persistence_attacker.create_deploy_key(
         "test/repo", "Test Key", "/tmp/test_key.pem"
